% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{health}
\alias{health}
\title{Synthetic stated preference data for healthcare interventions}
\format{
A data frame in wide format with the following columns:
\describe{
\item{id}{individual ID.}
\item{taskid}{choice task ID.}
\item{type.A, type.B, type.C}{type of the medical intervention for
alternatives A, B, and C (\code{1} = preventive, \code{2} = curative).}
\item{timeSpan.A, timeSpan.B, timeSpan.C}{time span until symptoms appear,
(\code{1} = after 20 years, \code{2} = after 5 years, \code{3} = within a year). }
\item{effectivenss.A, effectivenss.B, effectivenss.C}{the success probability (i.e. effectiveness) of the intervention
(\code{1} = 33\% effective, \code{2} = 66\% effective, \code{3} = 100\% effective).}
\item{adverseEffects.A, adverseEffects.B, adverseEffects.C}{the frequency of adverse effects
(\code{1} = often, \code{2} = rarely, \code{3} = never).}
\item{severity.A, severity.B, severity.C}{severity of illness
(\code{1} = not severe, \code{2} = severe, \code{3} = lethal).}
\item{lifestyle.A, lifestyle.B, lifestyle.C}{responsibility of patient's lifestyle for causing disease,
(\code{1} = fully, \code{2} = partly, \code{3} = not at all).}
\item{ageGroup.A, ageGroup.B, ageGroup.C}{age at onset of illness
(\code{1} = aged 80-90, \code{2} = aged 60-70, \code{3} = aged 40-50,
\code{4} = aged 20-30, \code{5} = aged 0-10).}
\item{rank.A, rank.B, rank.C}{reimbursement priority ranking for alternatives A, B, and C (lower values indicate higher priority).}
\item{best.A, best.B, best.C}{an indicator variable representing whether intervention(s) are considered for reimbursement (\code{1} = yes, \code{2} = no).}
}
}
\usage{
data("health", package = "groltir")
}
\description{
A generated dataset \code{health} containing simulated choices and priority rankings
for medical interventions with different attribute, allowing for ties.

The synthetic data includes three alternatives (A, B, and C), each representing a distinct medical
intervention. The profile of each intervention is a combination of seven attributes:
the type, the success probability (i.e. effectiveness), and the frequency of adverse effects,
of the intervention; severity of illness; patient responsibility for causing disease;
time span until symptoms appear; age at onset of illness.

We simulate a scenario with 100 individuals, each receiving 5 choice tasks.
For each task, a priority ranking is generated for the interventions, where a lower value in rank indicates a higher priority.
For instance, rank 1 represents the highest priority, rank 2 a slightly lower priority, with larger values indicating progressively lower priorities.
Additionally, choice data is generated to indicate which intervention(s) is/are preferred to be reimbursed,
allowing for multiple best choices in cases where respondents have no strict preference.
}
\keyword{datasets}
