% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirlogit.R
\name{InformationCriteria}
\alias{InformationCriteria}
\alias{AIC.tirlogit}
\alias{BIC.tirlogit}
\title{Information Criteria for the tirlogit model}
\usage{
\method{AIC}{tirlogit}(object, ...)

\method{BIC}{tirlogit}(object, ...)
}
\arguments{
\item{object}{a fitted model of class \code{tirlogit}.}

\item{...}{additional arguments (currently not used).}
}
\value{
For \code{AIC.tirlogit}, the AIC value. For \code{BIC.tirlogit}, the BIC value.
}
\description{
Calculates the Akaike Information Criterion (AIC)
and Bayesian Information Criterion (BIC) for an object of class \code{tirlogit},
based on the log-likelihood and the number of parameters.
}
\examples{
data("mode_wide", package = "tirlogit")
model_modeChoices<- tirlogit(multiBest ~ cost | dist + income , data = mode_wide,
                            intercept = FALSE, reflevel = "car",
                            ties = TRUE, type = "best", shape = "wide",
                            indexes = c("case", "taskid"), alts.names = "alt")
summary(model_modeChoices)
AIC(model_modeChoices)
BIC(model_modeChoices)
}
