% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRtest.R
\name{LRtest}
\alias{LRtest}
\title{Likelihood ratio test for \code{tirlogit} models}
\usage{
LRtest(model1, model2)
}
\arguments{
\item{model1}{a fitted model of class \code{tirlogit}.}

\item{model2}{a fitted model of class \code{tirlogit}.}
}
\value{
Summary of the likelihood ratio test, showing degrees of freedom,
log-likelihood values, chi-square statistic, and p-value.
}
\description{
The \code{LRtest} function performs a likelihood ratio tests
for two nested \code{tirlogit} models, to check whether
one model provides a significantly
better fit to the data than the other model.
}
\examples{
### Conduct a likelihood ratio test for nested models ###
data(mode_long, package = "tirlogit")
model1 <- tirlogit(rank ~ cost, data = mode_long, intercept = FALSE, reflevel = "car",
                   type = "rank", ties = TRUE, shape = "long", alts.names = "alt",
                   indexes = c("case", "taskid"))
model2 <- tirlogit(rank ~ cost + freq, data = mode_long, intercept = FALSE, reflevel = "car",
                   type = "rank", ties = TRUE, shape = "long", alts.names = "alt",
                   indexes = c("case", "taskid"))
LRtest(model1, model2)
}
