% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mode_long}
\alias{mode_long}
\title{Synthetic stated preference data for transportation modes}
\format{
A data frame in long format with the following columns:
\describe{
\item{case}{individual ID.}
\item{taskid}{choice task ID.}
\item{alt}{name of the alternative (i.e. transportation mode, one of train, car, bus and air).}
\item{dist}{choice set specific variable, the distance of the trip.}
\item{cost}{alternative specific variable, monetary cost associated with the transportation mode.}
\item{ivt}{alternative specific variable, in-vehicle travel time.}
\item{ovt}{alternative specific variable, out-of-vehicle travel time.}
\item{freq}{alternative specific variable, frequency of availability for the transportation mode.}
\item{income}{choice set specific variable, household income.}
\item{urban}{choice set specific variable, a dummy variable indicating whether the trip involves a large city as either
the origin or the destination (\code{1} = yes, \code{0} = no).}
\item{best}{a dummy variable indicating whether the alternative is selected as the best alternative; only one alternative can be selected as the best
in each choice task (\code{1} = most preferred, \code{0} = otherwise).}
\item{multiBest}{a dummy variable indicating whether the alternative is selected as the best alternative; multiple best choices can be given (\code{1} = most preferred, \code{0} = otherwise).}
\item{rankUnique}{preference ranking of the alternative; each alternative is assigned a unique rank; rank 1 for the most preferred alternative.}
\item{rank}{preference ranking of the alternative; ties are allowed; rank 1 for the most preferred alternative(s).}
\item{BestWorst}{an indicator variable representing whether the alternative is selected as the most preferred or least preferred alternative; multiple best and worst choices are allowed (\code{1} = most preferred, \code{-1} = least preferred, \code{0} = otherwise).}
}
}
\source{
This dataset is generated based on the \code{ModeCanada} dataset from the \code{R} package \code{mlogit}.
}
\usage{
data(mode_long, package = "groltir")
}
\description{
A generated dataset containing simulated choice data
for transportation modes across multiple choice tasks.
The dataset is structured in long format, where each row represents a specific transportation alternative
(e.g., train, air, bus, and car) within a choice task for an individual. Each alternative is characterized
by a set of attributes, including cost, in-vehicle and out-of-vehicle time, and frequency.

The data includes generated responses across five different types of choice tasks types, with responses in specific variables:
\itemize{
\item \strong{Best choice (\code{best})}: Indicate the most preferred alternative in each task, allowing only a single most preferred choice per task.
\item \strong{Best choice(s) (\code{multiBest})}: Indicate the most preferred alternative in each task, allowing for multiple best choices.
\item \strong{Ranking not allowing for ties (\code{rankUnique})}: Provide a preference ranking for each alternative within a task, without allowing for ties.
\item \strong{Ranking allowing for ties (\code{rank})}: Provide a preference ranking that allows ties, where lower values indicate higher preference. Tied alternatives share the same rank.
\item \strong{Best-worst choice(s) (\code{BestWorst})}: Indicate the most preferred and the least preferred alternative in each task.
}
}
\references{
Croissant, Y. (2020). \code{mlogit}: Multinomial Logit Models. R package version 1.1-1.
Available from: \url{https://CRAN.R-project.org/package=mlogit}.
}
\keyword{datasets}
