% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mode_wide}
\alias{mode_wide}
\title{Synthetic stated preference data for transportation modes}
\format{
A data frame in wide format.
}
\usage{
data(mode_wide, package = "groltir")
}
\description{
A generated dataset containing simulated choice data
for transportation modes across multiple choice tasks, in a wide format. This dataset is a wide-format
version of \link{mode_long}. Each row corresponds to a choice task.
It includes choice set specific variables (\code{dist}, \code{income}, and \code{urban}),
alternative specific attributes for each mode (\code{cost}, \code{freq}, \code{ivt}, and \code{ovt}),
and simulated preference data across various scenarios in separate columns.
For detailed descriptions of the dataset, see \link{mode_long}.
Additionally, this dataset also includes character string representations of each preference response.
}
\keyword{datasets}
