% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Encode categorical variables}
\usage{
recode(data, name, ctype, reflevel, ...)
}
\arguments{
\item{data}{a data frame containing the variables to be encoded.
The data frame can be in either long or wide format. Variables can be of type numeric or character.}

\item{name}{a character vector of variable names to encode.}

\item{ctype}{a character vector specifying the encoding type for each variable:
"C" for continuous (no encoding), "D" for dummy coding, and "E" for effect coding.}

\item{reflevel}{a character vector specifying the reference levels for encoding.
For variables of type character, the default \code{reflevel} is the alphabetically first level.
For numeric variables, the default \code{reflevel} is the level with smallest numeric value.}

\item{...}{additional arguments (currently not used).}
}
\value{
A data frame with the encoded variables based on the specified encoding types.
}
\description{
The function is used to encode categorical variables in a data frame, based on the specified encoding type.
It supports three encoding types: continuous, dummy, and effect coding.
The input data can be in either long or wide format.
}
\examples{
data(health, package = "groltir")
name <- c("type", "timeSpan", "effectivenss", "adverseEffects")
ctype <- c("D", "E", "C", "D")
reflevel <- c(1, 1, NA, 2)
health_coded <- recode(data = health, name, ctype, reflevel) # data in wide format
head(health_coded)
}
