% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirlogit.R
\name{summary.tirlogit}
\alias{summary.tirlogit}
\title{Summary method for tirlogit model objects.}
\usage{
\method{summary}{tirlogit}(object, ...)
}
\arguments{
\item{object}{a fitted model of class \code{tirlogit}.}

\item{...}{additional arguments (currently not used).}
}
\value{
Estimation results of the fitted tirlogit model, including:
\itemize{
\item The model call, showing how the \code{tirlogit} model was specified.
\item A table that includes the estimates of the coefficients, and the corresponding standard errors, z-values, p-values,
and significance codes.
\item The log-likelihood of the fitted model.
\item McFadden R^2 (only relevant if \code{Intercept = TRUE}).
\item Likelihood ratio test of the fitted model (vs. null model, only relevant if \code{Intercept = TRUE}).
}
}
\description{
Provides a summary of the estimation results of the fitted \code{tirlogit} model.
}
