% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirlogit.R
\name{tirlogit}
\alias{tirlogit}
\title{Analyze various types of discrete choice data with the Generalized Rank-Ordered Logit Model}
\usage{
tirlogit(
  formula,
  data,
  intercept = FALSE,
  reflevel = NULL,
  type = "rank",
  shape = "wide",
  alts.names = NULL,
  indexes = NULL,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{data}{an ordinary \code{data.frame} containing the variables in the model.}

\item{intercept}{logical; whether the alternative-specific intercepts are included in the model.}

\item{reflevel}{the reference level (the base alternative) for the utilities when estimating the model
(i.e. the one for which the coefficients of choice set specific covariates are set to zero);
The default \code{reflevel} is the alternative whose name appears first when sorted alphabetically from A to Z.}

\item{type}{a character string specifying the type of the input choice data; "\code{rank}" for full (complete) rankings,
"\code{best}" for best choice(s), "\code{BW}" for best-worst choice(s), "\code{string}" for general rankings in string format.}

\item{shape}{a character string specifying the shape of the input \code{data}, either \code{"long"} or \code{"wide"}.
By default, \code{shape = "wide"}.}

\item{alts.names}{only relevant if \code{shape = "long"}; a character string specifying
the name of the column that contains the names of the alternatives.}

\item{indexes}{a character vector specifying the names of
the columns that contain the individual IDs and the choice task IDs.
\code{indexes} can be omitted if \code{shape = "wide"}.}

\item{...}{additional arguments (currently not used).}
}
\value{
An object of class \code{tirlogit} containing model fit results including the estimated parameters,
log-likelihood, the variance-covariance matrix of the estimated parameters, and the convergence status.
}
\description{
The \code{tirlogit} function is used to fit the generalized rank-ordered logit model to analyze
various types of discrete choice data with and without ties, including the best choice(s),
best-worst choice(s), full ranking data with and without ties, etc.

The function supports a three-part formula syntax similar to that of the \code{mlogit} package, which includes:
\itemize{
\item alternative specific variables with generic coefficients
\item choice set specific or individual specific variables with specific coefficients
\item alternative specific variables with alternative specific coefficients.
}
}
