% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirlogit.R
\name{wtp}
\alias{wtp}
\title{Compute Willingness-to-Pay (WTP) Estimates for a tirlogit Model}
\usage{
wtp(model, var.names, wrt)
}
\arguments{
\item{model}{a fitted model of class \code{tirlogit}.}

\item{var.names}{a character vector specifying one or more variable names for
which the WTP will be computed.}

\item{wrt}{a single character string specifying the name of the cost variable
(the variable with respect to which the WTP is calculated).}
}
\value{
A table summarizing the WTP estimates, including the estimate, standard error,
t-statistic, p-value, and significance codes for each specified variable.
}
\description{
Computes the willingness-to-pay (WTP) estimates for specified variables
relative to a chosen cost variable in a fitted \code{tirlogit} model.
}
\examples{
mode_long$time <- mode_long$ivt + mode_long$ovt
model0 <- tirlogit(formula = best ~ cost + freq + time, data = mode_long, intercept = FALSE,
                   reflevel = NULL, type = "best", shape = "long",
                   indexes = c("case", "taskid"),
                   alts.names = "alt")
summary(model0)
wtp(model0, var.names = "freq" , wrt = "cost") # Compute WTP for a single variable
wtp(model0, var.names = c("freq", "time") , wrt = "cost") # Compute WTP for multiple variables
}
